+++ CHANGELOG.md
@@ -4,4 +4,5 @@ Defaults to `true`
 - Internal code refactoring
 - Add optional field, `"game_types"`, to group entries.
 Must be a list of game modes to check for. Ignored if it isn't 
-declared. 
+declared.
+- Add simple version of ban command